atom defaultValues refactor
