+++ core-ui/src/state/clusterConfigAtom.ts
@@ -3,13 +3,10 @@ import { atom, RecoilState } from 'recoil';
 export type ClusterConfigState = {
   requiresCA: boolean;
   storage: 'localStorage' | 'sessionStorage' | string;
-} | null;
-
-const defaultValue: ClusterConfigState = null;
+};
 
 export const clusterConfigState: RecoilState<ClusterConfigState> = atom<
   ClusterConfigState
 >({
   key: 'clusterConfigState',
-  default: defaultValue,
 });
