+++ core-ui/src/state/extResourcesAtom.ts
@@ -1,13 +1,10 @@
 import { atom, RecoilState } from 'recoil';
 import { ExtResource } from './types';
 
-type ExtResourcesAtom = ExtResource[] | null;
-
-const defaultValue: ExtResourcesAtom = null;
+type ExtResourcesAtom = ExtResource[];
 
 export const extResourcesState: RecoilState<ExtResourcesAtom> = atom<
   ExtResourcesAtom
 >({
   key: 'extResourcesState',
-  default: defaultValue,
 });
