+++ core-ui/src/state/navigation/filters/shouldNodeBeVisible.ts
@@ -2,7 +2,7 @@ import { areNodeFeaturesEnabled } from './areNodeFeaturesEnabled';
 import { doesNodeResourceExist } from './doesNodeResourceExist';
 import { isNodeResourcePermitted } from './isNodeResourcePermitted';
 import { ConfigFeatureList, NavNode } from '../../types';
-import { PermissionSet } from '../../permissionSetsAtom';
+import { PermissionSet } from '../../permissionSetsSelector';
 
 type NavConfigSet = {
   configFeatures: ConfigFeatureList;
