+++ core-ui/src/state/openapi/openapiAtom.ts
@@ -5,9 +5,6 @@ type OpenapiState = {
   paths: Record<string, any>;
 } | null;
 
-const defaultValue = null;
-
 export const openapiState: RecoilState<OpenapiState> = atom<OpenapiState>({
   key: 'openapiState',
-  default: defaultValue,
 });
