+++ core-ui/src/state/resourceList/resourceListSelector.ts
@@ -3,7 +3,7 @@ import { extResourcesState } from '../extResourcesAtom';
 import { resources } from 'resources';
 import { partial } from 'lodash';
 import { NavNode } from '../types';
-import { configFeaturesState } from '../configFeaturesAtom';
+import { configFeaturesState } from '../configFeatures/configFeaturesAtom';
 import { mapBusolaResourceToNavNode } from './mapBusolaResourceToNavNode';
 import { mapExtResourceToNavNode } from './mapExtResourceToNavNode';
